requirements.txt
matplotlib==3.7.2
pandas==2.0.0
Pillow==9.5.0
Pillow==10.1.0
torch==2.0.0
torchsummary==1.5.1
torchvision==0.15.1
